Added status led configuration value in html and some error checks on loading config
Added new option in HTML peripheral configuration to add an optional or
whatever that signals that the relay is in on/off state. Also added a
flag that indicates the output is active low or active high.

Added a few more checks on loading the json file for existing keys,
maybe change the save function to only save values that are enabled not
all like now